cleanup and wire bind command
