+++ kcp/apiexport-catalog.kcp.dev.yaml
@@ -1,0 +1,9 @@
+apiVersion: apis.kcp.dev/v1alpha1
+kind: APIExport
+metadata:
+  creationTimestamp: null
+  name: catalog.kcp.dev
+spec:
+  latestResourceSchemas:
+  - v221005-87667ee.catalogentries.catalog.kcp.dev
+status: {}
